+++ lib/js_routes/generators/webpacker.rb
@@ -7,7 +7,7 @@ class JsRoutes::Generators::Webpacker < Rails::Generators::Base
   def create_webpack
     copy_file "initializer.rb", "config/initializers/js_routes.rb"
     copy_file "erb.js", "config/webpack/loaders/erb.js"
-    copy_file "routes.js.erb", "app/javascript/routes.js.erb"
+    copy_file "routes.js.erb", "#{Webpacker.config.source_path}/routes.js.erb"
     inject_into_file "config/webpack/environment.js", loader_content
     if path = application_js_path
       inject_into_file path, pack_content